// apps/bc-wallet-showcase-creator/schemas/character.ts
import { z } from "zod";

export const characterSchema = z.object({
  name: z.string().min(1, "Name is required"),
  role: z.string().min(1, "Role is required"),
  description: z.string().min(1, "Description is required").max(100, 'Description must be less than 100 characters'),
  hidden: z.boolean().optional(),
  slug: z.string().optional()
});